+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - sdk(sdk): Add deposit simulation.
+- sdk(sdk): Add GLV deposit simulation.
 - cli: Add auto-creation support of multiple ALTs based on ALT limit for `alt extend` command.
 
 ### Changed